SensorGatewayService/requirements.txt
nats-py==2.12.0
paho-mqtt==2.1.0
python-dotenv==1.2.1
